# Integrate RowMatrix into the prover