--- a/prover/src/constraints/composition_poly.rs
+++ b/prover/src/constraints/composition_poly.rs
@@ -3,7 +3,7 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use super::{Matrix, StarkDomain};
+use super::{ColMatrix, StarkDomain};
 use math::{polynom, FieldElement, StarkField};
 use utils::{collections::Vec, uninit_vector};
 
@@ -13,7 +13,7 @@ use utils::{collections::Vec, uninit_vector};
 /// to trace_length. Thus, for example, if the composition polynomial has degree 2N - 1, where N
 /// is the trace length, it will be stored as two columns of size N (each of degree N - 1).
 pub struct CompositionPoly<E: FieldElement> {
-    data: Matrix<E>,
+    data: ColMatrix<E>,
 }
 
 impl<E: FieldElement> CompositionPoly<E> {
@@ -43,7 +43,7 @@ impl<E: FieldElement> CompositionPoly<E> {
         let polys = transpose(coefficients, num_columns);
 
         CompositionPoly {
-            data: Matrix::new(polys),
+            data: ColMatrix::new(polys),
         }
     }
 
@@ -71,7 +71,7 @@ impl<E: FieldElement> CompositionPoly<E> {
     // --------------------------------------------------------------------------------------------
     /// Evaluates the columns of the composition polynomial over the specified LDE domain and
     /// returns the result.
-    pub fn evaluate<B>(&self, domain: &StarkDomain<B>) -> Matrix<E>
+    pub fn evaluate<B>(&self, domain: &StarkDomain<B>) -> ColMatrix<E>
     where
         B: StarkField,
         E: FieldElement<BaseField = B>,