--- a/prover/src/constraints/mod.rs
+++ b/prover/src/constraints/mod.rs
@@ -3,7 +3,7 @@
 // This source code is licensed under the MIT license found in the
 // LICENSE file in the root directory of this source tree.
 
-use super::{ConstraintDivisor, Matrix, ProverError, StarkDomain};
+use super::{ColMatrix, ConstraintDivisor, ProverError, StarkDomain};
 
 mod boundary;
 use boundary::BoundaryConstraints;