--- a/prover/src/lib.rs
+++ b/prover/src/lib.rs
@@ -78,7 +78,7 @@ mod domain;
 pub use domain::StarkDomain;
 
 mod matrix;
-pub use matrix::{Matrix, RowMatrix};
+pub use matrix::{ColMatrix, RowMatrix};
 
 mod constraints;
 use constraints::{CompositionPoly, ConstraintCommitment, ConstraintEvaluator};
@@ -456,9 +456,9 @@ pub trait Prover {
     /// building a Merkle tree from the resulting hashes.
     fn build_trace_commitment<E>(
         &self,
-        trace: &Matrix<E>,
+        trace: &ColMatrix<E>,
         domain: &StarkDomain<Self::BaseField>,
-    ) -> (Matrix<E>, MerkleTree<Self::HashFn>, Matrix<E>)
+    ) -> (ColMatrix<E>, MerkleTree<Self::HashFn>, ColMatrix<E>)
     where
         E: FieldElement<BaseField = Self::BaseField>,
     {