… examples/src/rescue_raps/custom_trace_table.rs
… prover/src/trace/mod.rs
… prover/src/trace/poly_table.rs
… prover/src/trace/tests.rs
… prover/src/trace/trace_lde.rs
… prover/src/trace/trace_table.rs
… winterfell/src/lib.rs