This PR integrates `RowMatrix` struct into the prover. Specifically:

- Changes the return type of `build_trace_commitment()` function in the prover.
- Updates `TraceLde` struct to hold `RowMatrix`'s instead of `ColMatrix`'s.
- Propagated the above changes to other components to make sure everything compiles.

For small proofs, the effect of these changes is usually not noticeable. However, for large proofs (e.g., $2^{20}$ rows, dozens of columns) the speed up in LDE and Merkle Tree construction steps could be significant.