@Nashtare 2023-03-16
> However, for large proofs (e.g., $2^{20}$ rows, dozens of columns) the speed up in LDE and Merkle Tree construction steps could be significant.

Just to complement on this point, benchmarking fib-small (2 columns) and lamport-threshold (28 columns), the two extremities of the spectrum among the examples, I am getting on my machine:

- a slowdown of 140% for the LDE extension step on the fib-small example (2^20 rows, blowup = 16), from 1406ms to 3375ms.
- a speedup of 48% for the LDE extension step on the lamport-t example (2^19 rows, blowup = 16), from 16672ms to 8803ms.

I don't think the slowdown for narrow traces is really an issue as practically speaking we will most likely always optimize proving time (and amortize verifier if needed through recursive composition in a near future), but thought it was worth mentioning in case some people notice a performance downgrade on the existing examples (most of which having very few columns).