@irakliyk 2023-03-17
Thank you for the review!

> a slowdown of 140% for the LDE extension step on the fib-small example (2^20 rows, blowup = 16), from 1406ms to 3375ms.

Ah - interesting! I have't tested it on such narrow traces - but I think with the current settings it is to be expected. Basically, since segment width is set to 8, we actually run NTTs on 8 columns instead of 2. I wonder if this degradation can be eliminated by setting segment width to 2.

> a speedup of 48% for the LDE extension step on the lamport-t example (2^19 rows, blowup = 16), from 16672ms to 8803ms.

What machine are you running this on? I've tested on an M1 and on AMD Ryzen 9 5950X and there was no improvement in the former and about 3% degradation in the latter (though, changing segment width to 4 resulted in a slight improvement).

> I don't think the slowdown for narrow traces is really an issue

Yes, that's what I think too. In most practical situations we'll have more than just a few columns and if needed, we might be able to get back to the original performance by adjusting segment width.