@Nashtare 2023-03-19
> What machine are you running this on? I've tested on an M1 and on AMD Ryzen 9 5950X and there was no improvement in the former and about 3% degradation in the latter (though, changing segment width to 4 resulted in a slight improvement).

Same as before when I'm commenting about benchmarks, Ubuntu 22.04 Intel® Core™ i7-9750H CPU @ 2.60GHz × 12 CPUs. (although as we've discussed in the past, there may be weird results sometimes with Linux and/or x86_64, as with the RP64 benchmarks for instance)..